Updated includes & library functions.

+++ Wrappers/Font.cpp
@@ -1,11 +1,11 @@
 ﻿#include <stdexcept>
 
-#include <SDL_ttf.h>
+#include <SDL3_ttf/SDL_ttf.h>
 
 #include "Font.h"
 
 
-bae::Font::Font(const std::string& fullPath, unsigned int size) : m_font(nullptr)
+bae::Font::Font(const std::string& fullPath, float size) : m_font(nullptr)
 {
 	m_font = TTF_OpenFont(fullPath.c_str(), size);
 	if (m_font == nullptr)
